Update README.md
The Student Record Management System is a console-based application developed in C++ as part of the Programming Fundamentals course. The project provides a simple and efficient way to manage student records such as roll number, name, and marks. It replaces manual record-keeping with a computerized solution that ensures accuracy, efficiency, and data persistence.